feat: Add last message and last message time to Conversation model
This commit adds the `lastMessage` and `lastMessageAt` properties to the `Conversation` model in the `chatModel.kt` file. These properties represent the last message content and the timestamp of the last message in a conversation. This modification enhances the functionality of the application by allowing users to see the most recent message in their conversations.