feat: Add openNotifications function to Home.kt
This commit adds the `openNotifications` function to the `Home.kt` file. The function is responsible for handling the click event on the notification icon in the home screen. It checks if the user is authenticated and then navigates to the Notifications activity if they are, or to the LoginWa activity if they are not. This addition enhances the functionality of the application by allowing users to easily access their notifications.